requirements.txt
requests
feedparser
openai
schedule
scikit-learn
python-dateutil
playwright
